requirements.txt
numpy
scikit-image
opencv-python
pyfingerprint
RPi.GPIO